backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
elasticsearch==8.14.0
pysolr==3.10.0b2
